requirements.txt
cffi>=1.16.0
click>=8.1.7
colorama>=0.4.6
curl-cffi>=0.5.10
duckduckgo_search>=4.2
lxml>=5.1.0
pycparser>=2.21
python-dateutil>=2.8.2
six>=1.16.0

marvin~=2.1.2
langchain~=0.1.0